+++ package.json
@@ -25,7 +25,7 @@
     "react": "17.0.1",
     "react-dom": "17.0.1",
     "react-icons": "4.2.0",
-    "redoc": "2.0.0-rc.53",
+    "redoc": "2.0.0-rc.56",
     "sass": "1.32.12",
     "selic": "1.0.0",
     "styled-components": "5.3.0",
