+++ pages/api/cptec/v1/ondas/[cityCode]/index.js
@@ -1,5 +1,5 @@
 import app from '@/app';
-import NotFoundError from '@/errors/not-found';
+import NotFoundError from '@/errors/NotFoundError';
 
 import { getSwellData } from '@/services/cptec';
